move 2D file tests into main test file
to avoid missing reference figures
