+++ tests/_base_testing.py
@@ -1,4 +1,5 @@
 import os
+import unittest
 import sys
 import subprocess as spr
 from unittest import TestCase
@@ -118,3 +119,25 @@ class PsyPlotTestCase(TestCase):
                                        err_msg=msg or '', **kwargs)
         except AssertionError as e:
             self.fail(e.message)
+
+
+class TestBase2D:
+    """Test :class:`psyplot.plotter.baseplotter.BasePlotter` class without time
+    and vertical dimension"""
+
+    def _label_test(self, key, label_func, has_time=False):
+        return super(TestBase2D, self)._label_test(
+            key, label_func, has_time=has_time)
+
+    def __getattribute__(self, attr):
+        """Hack to disable the creation of reference figures"""
+        ret = super().__getattribute__(attr)
+        if callable(ret) and attr.startswith("ref_"):
+
+            @unittest.skip("Reference figures are created from the base class")
+            def skip_func(self):
+                pass
+
+            skip_func.__name__ = attr
+            return skip_func()
+        return ret
