+++ tests/test_base.py
@@ -141,26 +141,11 @@ class BasePlotterTest(bt.PsyPlotTestCase):
             self.assertGreaterEqual(data[data > 250].max(), 251)
 
 
-class TestBase2D:
+class BasePlotterTest2D(bt.TestBase2D, BasePlotterTest):
     """Test :class:`psyplot.plotter.baseplotter.BasePlotter` class without time
     and vertical dimension"""
 
-    def _label_test(self, key, label_func, has_time=False):
-        return super(TestBase2D, self)._label_test(
-            key, label_func, has_time=has_time)
-
-    def __getattribute__(self, attr):
-        """Hack to disable the creation of reference figures"""
-        ret = super().__getattribute__(attr)
-        if callable(ret) and attr.startswith("ref_"):
-
-            @unittest.skip("Reference figures are created from the base class")
-            def skip_func(self):
-                pass
-
-            skip_func.__name__ = attr
-            return skip_func()
-        return ret
+    var = 't2m_2d'
 
 
 if __name__ == '__main__':
