+++ tests/test_plotters_combinedplotter_circumpolar.py
@@ -154,5 +154,18 @@ class CircumpolarCombinedPlotterTest(tpc.CombinedPlotterTest):
         pass
 
 
+class CombinedPlotterTest2D(bt.TestBase2D, tpc.CombinedPlotterTest):
+    """Test :class:`psyplot.plotter.maps.CombinedPlotter` class without time and
+    vertical dimension"""
+
+    var = ['t2m', ['u_2d', 'v_2d']]
+
+    def _label_test(self, key, label_func, has_time=None):
+        if has_time is None:
+            has_time = not bool(self.vector_mode)
+        tpc.CombinedPlotterTest._label_test(
+            self, key, label_func, has_time=has_time)
+
+
 if __name__ == '__main__':
     unittest.main()
